requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
intel-cmplr-lib-ur==2025.0.0
intel-openmp==2025.0.0
Jinja2==3.1.3
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.18
langchain-openai==0.2.8
langchain-xai==0.1.0
langgraph==0.2.48
langgraph-checkpoint==2.0.4
langgraph-sdk==0.1.36
langsmith==0.1.143
MarkupSafe==2.1.5
mkl==2025.0.0
mkl-service==2.4.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.3
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==10.2.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tbb==2022.0.0
tcmlib==1.2.0
tenacity==9.0.0
tiktoken==0.8.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
umf==0.9.0
urllib3==2.2.3
wheel==0.44.0
yarl==1.17.1
